% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RisingAshes-package.R
\docType{data}
\name{Devecey}
\alias{Devecey}
\title{Devecey}
\description{
Forest field trial consisting on 23 open-pollinated common ash (Fraxinus
excelsior) maternal progenies from 3 North-Eastern French provenances.
}
\details{
Variables are:
\describe{
 \item{Seq}{tree id}
 \item{X, Y}{coordinates (row/cols, spaced 4 m x 4 m)}
 \item{BLC}{(randomized incomplete) block}
 \item{SBLC}{sub-block}
 \item{FAM}{OP Family (23 levels)}
 \item{PROV}{Provenance (3 levels)}
 \item{BF98}{Bud-Flush precocity ranking (1 - late; 5 - early)}
 \item{CD10..CD14}{Crown Dieback proportion (classes 0, 0.05, 0.30, 0.65, 0.90, 1) in 2010..2014}
 \item{CL12..CL13}{Collar Lesion proportion (continuous measure from 0 to 1) in 2012 and 2013}
 \item{BC12..BC13}{Basal Circumference (in cm) in 2012 and 2013}
}
}
\references{
F. Muñoz, B. Marçais, J. Dufour and A. Dowkiw (2016). Rising Out
  of the Ashes: Additive Genetic Variation for Crown and Collar Resistance to
  Hymenoscyphus fraxineus in Fraxinus excelsior. Phytopathology.
  http://dx.doi.org/10.1094/PHYTO-11-15-0284-R

  Preprint bioRXiv http://dx.doi.org/10.1101/031393
}

