% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Devecey_spde}
\alias{Devecey_spde}
\title{SPDE structure for the Devecey dataset}
\usage{
Devecey_spde(coord = Devecey[, c("X", "Y")])
}
\value{
A list with the observation locations (loc) a prediction mesh and an
  SPDE object which includes the prior distributions for its hyperparameters
}
\description{
SPDE structure for the Devecey dataset
}

