% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fam_additive}
\alias{fam_additive}
\title{Additive-genetic matrices for a family structure}
\usage{
fam_additive(id, fam)
}
\arguments{
\item{id}{numeric vector of tree ids}

\item{fam}{factor with family codes}
}
\value{
a list with a \code{recoding} of the original \code{id} codes, the
  Inverse relationship matrix (\code{Ainv}) in sparse format, and the
  family-wise sum-to-zero \code{constraint.matrix} for use in conjuntion with
  an explicit family effect
}
\description{
Computes the relevant matrices to set-up an additive-genetic effect given a
family structure
}

