% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{famped}
\alias{famped}
\title{Pedigree for a family structure}
\usage{
famped(id, fam)
}
\arguments{
\item{id}{numeric vector of tree ids}

\item{fam}{factor with family codes}
}
\value{
A three-column data.frame with numeric codes for the individual trees
  and their progenitors
}
\description{
Compute the pedigree from tree id and family codes
}
\details{
Note that renumbering might take place.
}

