% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inv.logit}
\alias{inv.logit}
\title{Inverse logit function}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{value(s) to be transformed

\deqn{p = exp(x)/(1 + exp(x))}}
}
\description{
Compute the inverse logit
}

