% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pedigree}
\alias{Ainverse}
\alias{RenamePedigree}
\alias{SortPedigree}
\alias{compute_Ainverse}
\alias{pedigree}
\title{Pedigree functions}
\usage{
compute_Ainverse(pedigree)

SortPedigree(pedigree)

RenamePedigree(pedigree)

Ainverse(pedigree)
}
\arguments{
\item{pedigree}{3-col numerical matrix}
}
\description{
These functions are borrowed from the package \code{AnimalINLA}
}
\section{Functions}{
\itemize{
\item \code{compute_Ainverse}: Check, recode and compute the inverse relationship matrix

\item \code{SortPedigree}: Sort a pedigree chronologically

\item \code{RenamePedigree}: Recode a pedigree

\item \code{Ainverse}: Compute the inverse relationship matrix
}}
\references{
http://www.r-inla.org/related-projects/animalinla
}

