% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RisingAshes-package.R
\docType{package}
\name{RisingAshes}
\alias{RisingAshes}
\alias{RisingAshes-package}
\title{RisingAshes}
\description{
Data and code to reproduce the results in Muñoz et al. (2016), which analyzes
the genetic variation in susceptibility to a pathogen using a spatio-temporal
hierarchical Bayesian model.
}
\details{
This package contains three vignettes with all the code necessary to
reproduce the results from the paper (and more results that did not make it
to the paper). You can browse the code with \code{browseVignettes(package =
'RisingAshes')} and work with the data with \code{data(Devecey) }

Since the computations can take several hours, all the expensive results are
lazy-loaded with the package, so that you can explore the results without the
need of recomputing everything. The code for producing the results is
included in the vignettes, but is not executed by default at compilation
time. For reproducing the results you might want to setup the computation on
a server.
}
\references{
F. Muñoz, B. Marçais, J. Dufour and A. Dowkiw (2016). Rising Out
  of the Ashes: Additive Genetic Variation for Crown and Collar Resistance to
  Hymenoscyphus fraxineus in Fraxinus excelsior. Phytopathology.
  http://dx.doi.org/10.1094/PHYTO-11-15-0284-R

  Preprint bioRXiv http://dx.doi.org/10.1101/031393
}

