% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{shift_marginal}
\alias{shift_marginal}
\title{Perform a shift on a (list of) marginals}
\usage{
shift_marginal(mar, mu)
}
\arguments{
\item{mar}{either a single marginal or a list of marginals}

\item{mu}{numeric. Constant to be added.

\code{INLA}'s marginals are two-column numeric matrices with abscissas and
ordinates. This function adds a constant \code{mu} to the ordinates (i.e.,
second column).}
}
\description{
Add a constant to a \code{INLA}'s marginal distribution or list of marginals.
}

