% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{suffix}
\alias{suffix}
\title{Include a suffix to variable names}
\usage{
suffix(x, suffix, makeNA = FALSE)
}
\arguments{
\item{x}{a data.frame}

\item{suffix}{character string to be appended to variable names}

\item{makeNA}{logical. If \code{TRUE}, returns the suffixed data.frame filled
with \code{NA}.}
}
\description{
Used with \code{\link[INLA]{inla.stack}} for building effect stacks.
}

